requirements.txt
numpy
torch
datetime
matplotlib
termcolor
pyyaml